requirements.txt
pillow
requests
numpy
rembg
tkinterdnd2
sv-ttk
